pyproject.toml
[tool.poetry]
name = "llamaindex"
version = "0.1.0"
description = ""
authors = ["Dylan Sclater <dylancsclater@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.31.0"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"
llama-index = "^0.10.18"
openai = "^1.13.3"
pinecone-client = "^3.1.0"
llama-index-vector-stores-pinecone = "^0.1.4"
nbconvert = "^7.16.2"
pypdf = "^4.1.0"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


